+++ docs/changelog.md
@@ -4,10 +4,13 @@
 
 ### Changed
 
-- [promql/series](promql/../checks/promql/series.md) check now supports more flexible
+- [promql/series](checks/promql/series.md) check now supports more flexible
   `# pint disable promql/series(...)` comments.
   Adding a comment `# pint disable promql/series({cluster="dev"})` will disable this check
   for any metric selector with `cluster="dev"` matcher.
+- [query/cost](checks/query/cost.md) check will now calculate how much Prometheus memory
+  will be needed for storing results of given query.
+  `bytesPerSample` option that was previously used to calculate this was removed.
 
 ## v0.27.0